Moved all state logic and helper functions to useDropdown hook.
