+++ headless-dropdown-component/src/Components/Dropdown/hooks/useDropdown.ts
@@ -1,6 +1,90 @@
 import { useState } from "react";
-import { DropdownItem } from "../DropdownMenu";
+import { DropdownItem } from "../types/DropdownTypes";
 
-const useDropdown = (items: DropdownItem[]) => {
+const useDropdown = (dropdownItems: DropdownItem[]) => {
   const [isOpen, setIsOpen] = useState(false);
+  const [selectedItem, setSelectedItem] = useState<DropdownItem | null>(null);
+  const [selectedIndex, setSelectedIndex] = useState<number>(-1);
+
+  const handleDropdownItemClick = (currentItem: DropdownItem) => {
+    setSelectedItem(currentItem);
+    setIsOpen(false);
+  };
+
+  const handleKeyDown = (e: React.KeyboardEvent) => {
+    switch (e.key) {
+      // Toggle menu and update trigger text if users made a selection
+      case "Enter":
+      case " ":
+        setIsOpen(!isOpen);
+        if (selectedIndex !== -1) {
+          setSelectedItem(dropdownItems[selectedIndex]);
+        }
+        break;
+
+      // Ascending traveral through list items 1-N. Wraps to item 1.
+      case "ArrowDown":
+      case "ArrowRight":
+        e.preventDefault();
+        e.stopPropagation();
+
+        if (isOpen !== true) {
+          setIsOpen(true);
+        }
+        if (
+          selectedIndex === -1 ||
+          selectedIndex === dropdownItems.length - 1
+        ) {
+          setSelectedIndex(0);
+        } else {
+          setSelectedIndex(selectedIndex + 1);
+        }
+        break;
+
+      // Descending traveral through list items N-1. Wraps to item N.
+      case "ArrowUp":
+      case "ArrowLeft":
+        e.preventDefault();
+        e.stopPropagation();
+
+        if (isOpen !== true) {
+          setIsOpen(true);
+        }
+
+        if (selectedIndex === -1) {
+          setSelectedIndex(0);
+        } else if (selectedIndex === 0) {
+          setSelectedIndex(dropdownItems.length - 1);
+        } else {
+          setSelectedIndex(selectedIndex - 1);
+        }
+        break;
+
+      // Close menu immediately
+      case "Escape":
+        if (isOpen) {
+          setIsOpen(false);
+        }
+        break;
+
+      // TODO: Add Home key to jump to item 1
+      // TODO: Add End key to jump to item N
+
+      default:
+        null;
+    }
+  };
+
+  const toggleDropdown = () => setIsOpen(!isOpen);
+
+  return {
+    isOpen,
+    selectedIndex,
+    selectedItem,
+    handleDropdownItemClick,
+    handleKeyDown,
+    toggleDropdown,
+  };
 };
+
+export default useDropdown;
